+++ src/tcp.h
@@ -57,8 +57,8 @@ int tcp_connect(tcp_t *tcp, char *hostname, int port, int secure,
 		char *local_if, char *message, unsigned io_timeout);
 void tcp_close(tcp_t *tcp);
 
-int tcp_read(tcp_t *tcp, void *buffer, int size);
-int tcp_write(tcp_t *tcp, void *buffer, int size);
+ssize_t tcp_read(tcp_t *tcp, void *buffer, int size);
+ssize_t tcp_write(tcp_t *tcp, void *buffer, int size);
 
 int get_if_ip(char *dst, size_t len, const char *iface);
